Now runs on WCF as the backend, instead of plain sockets.
git-svn-id: https://rhino-tools.svn.sourceforge.net/svnroot/rhino-tools/experiments/NMemcached@1461 079b0acf-d9fa-0310-9935-e5ade295c882

+++ NMemcached/Model/CachedItem.cs
@@ -6,7 +6,7 @@ namespace NMemcached.Model
 	{
 		private byte[] buffer;
 
-		public uint Flags;
+		public long Flags;
 		public string Key;
 		public long Timestamp;
